+++ egui/src/widgets/text_edit.rs
@@ -320,7 +320,7 @@ impl<'t> TextEdit<'t> {
         } else {
             Sense::hover()
         };
-        let response = ui.interact(rect, id, sense);
+        let mut response = ui.interact(rect, id, sense);
 
         if enabled {
             ui.memory().interested_in_kb_focus(id);
@@ -346,6 +346,7 @@ impl<'t> TextEdit<'t> {
                         primary: galley.from_ccursor(ccursorp.primary),
                         secondary: galley.from_ccursor(ccursorp.secondary),
                     });
+                    response.mark_changed();
                 } else if response.hovered() && ui.input().pointer.any_pressed() {
                     ui.memory().request_kb_focus(id);
                     if ui.input().modifiers.shift {
@@ -357,9 +358,11 @@ impl<'t> TextEdit<'t> {
                     } else {
                         state.cursorp = Some(CursorPair::one(cursor_at_pointer));
                     }
+                    response.mark_changed();
                 } else if ui.input().pointer.any_down() && response.is_pointer_button_down_on() {
                     if let Some(cursorp) = &mut state.cursorp {
                         cursorp.primary = cursor_at_pointer;
+                        response.mark_changed();
                     }
                 }
             }
@@ -484,6 +487,8 @@ impl<'t> TextEdit<'t> {
                 };
 
                 if let Some(new_ccursorp) = did_mutate_text {
+                    response.mark_changed();
+
                     // Layout again to avoid frame delay, and to keep `text` and `galley` in sync.
                     let font = &ui.fonts()[text_style];
                     galley = if multiline {